@@ -16,10 +16,13 @@ class NoisyWorkflow < ActionMailer::Base
   def request_fact_check(action)
     @edition = action.edition
     fact_check_address = @edition.fact_check_email_address
-    mail(to: action.email_addresses, reply_to: fact_check_address,
+    mail(
+      to: action.email_addresses,
+      reply_to: fact_check_address,
       from: "GOV.UK Editorial Team <#{fact_check_address}>",
-      subject: "‘[#{@edition.title}]’ GOV.UK preview of new edition") do |format|
-      format.text { render text: action.customised_message }
+      subject: "‘[#{@edition.title}]’ GOV.UK preview of new edition"
+    ) do |format|
+      format.text { render html: "<strong>HTML</strong> #{action.customised_message}" }
     end
   end
 
