@@ -1,5 +1,5 @@
-# Load the rails application
-require File.expand_path('../application', __FILE__)
+# Load the rails application.
+require_relative 'application'
 
-# Initialize the rails application
+# Initialize the rails application.
 Rails.application.initialize!
