@@ -6,6 +6,8 @@ class RepublishContentTest < ActiveSupport::TestCase
     @published_edition = FactoryGirl.create(:answer_edition, state: 'published')
     draft_artefact = FactoryGirl.create(:draft_artefact, kind: 'help_page', slug: 'help/me')
     @draft_edition = FactoryGirl.create(:help_page_edition, state: 'draft', panopticon_id: draft_artefact.id)
+
+    $stdout.stubs(puts: '')
   end
 
   context "#publishing_api:republish_content" do
@@ -13,12 +15,7 @@ class RepublishContentTest < ActiveSupport::TestCase
       RepublishWorker.expects(:perform_async).with(@published_edition.id.to_s)
       UpdateWorker.expects(:perform_async).with(@draft_edition.id.to_s)
 
-      # This is to prevent the rake task ouputting to the console
-      silence_stream(STDOUT) do
-        silence_stream(STDERR) do
-          Rake::Task['publishing_api:republish_content'].invoke
-        end
-      end
+      Rake::Task['publishing_api:republish_content'].invoke
     end
   end
 
@@ -26,12 +23,7 @@ class RepublishContentTest < ActiveSupport::TestCase
     should "only republish items of that format" do
       UpdateWorker.expects(:perform_async).with(@draft_edition.id.to_s)
 
-      # This is to prevent the rake task ouputting to the console
-      silence_stream(STDOUT) do
-        silence_stream(STDERR) do
-          Rake::Task['publishing_api:republish_by_format'].invoke('help_page')
-        end
-      end
+      Rake::Task['publishing_api:republish_by_format'].invoke('help_page')
     end
   end
 end
